Implementation of Threading Concept in java
A Thread is a very light-weighted process, or we can say the smallest part of the process that allows a program to operate more efficiently by running multiple tasks simultaneously. In order to perform complicated tasks in the background, we used the Thread concept in Java. All the tasks are executed without affecting the main program.