implementation of Multiplication of matrix in Java
Matrix Multiplication in Java Obtaining a single matrix from the entries of two matrices by using a binary operation is known as Matrix multiplication. In simpler terms, if two matrices matrix1 and matrix2 of order a*b and b*c are multiplied, the matrix obtained is of the order a*c.